Reverse the order of yip-6 and yip-7
