+++ dao-v2/yip-06/yip-06-michelson.tz
@@ -1,1 +1,1 @@
-{ DROP; NIL operation; PUSH address "KT1LrEJsaTR5vMdwjvASTtFPUbk2wnX3P166"; CONTRACT %set_contracts (pair (pair (pair address address) (pair address address)) (pair address address)); IF_NONE { PUSH int 215; FAILWITH } {}; PUSH mutez 0; PUSH (pair (pair (pair address address) (pair address address)) (pair address address)) (Pair (Pair (Pair "KT1SjXiUX63QvdNMcM2m492f7kuf8JxXRLp4" "KT1LpDGYxbpabK3A6rqpVW4TSwVjwg9Zjp7K") (Pair "KT1Ge5usHwAAH12PDxwP8FFYMdbMbXSRXSz4" "KT1PJkeVGVJAH4LRjWRbLU39D9EY7GkTfibQ")) (Pair "KT1TYSXHmkGu7QSsiQKCahWpKtG3JtFgN2kf" "KT1FPmpucXoiX7ZLahj1V1E5tRah1XvcnkZB")); TRANSFER_TOKENS; CONS; PUSH address "KT1UJBvm4hv11Uvu6r4c8zE5K2EfmwiRVgsm"; CONTRACT %setRewardRecipient address; IF_NONE { PUSH int 229; FAILWITH } {}; PUSH mutez 0; PUSH address "KT1FPmpucXoiX7ZLahj1V1E5tRah1XvcnkZB"; TRANSFER_TOKENS; CONS; PUSH address "KT1XvH5f2ja2jzdDbv6rxPmecZFU7s3obquN"; CONTRACT %setRewardRecipient address; IF_NONE { PUSH int 229; FAILWITH } {}; PUSH mutez 0; PUSH address "KT1FPmpucXoiX7ZLahj1V1E5tRah1XvcnkZB"; TRANSFER_TOKENS; CONS; PUSH address "KT1Xbx9pykNd38zag4yZvnmdSNBknmCETvQV"; CONTRACT %setRewardRecipient address; IF_NONE { PUSH int 229; FAILWITH } {}; PUSH mutez 0; PUSH address "KT1FPmpucXoiX7ZLahj1V1E5tRah1XvcnkZB"; TRANSFER_TOKENS; CONS; PUSH address "KT1AVbWyM8E7DptyBCu4B5J5B7Nswkq7Skc6"; CONTRACT %setRewardRecipient address; IF_NONE { PUSH int 229; FAILWITH } {}; PUSH mutez 0; PUSH address "KT1FPmpucXoiX7ZLahj1V1E5tRah1XvcnkZB"; TRANSFER_TOKENS; CONS; PUSH address "KT1STLQKxiRtAh1e7DZhu1xUTAJ7KLpV9Rru"; CONTRACT %setRewardRecipient address; IF_NONE { PUSH int 229; FAILWITH } {}; PUSH mutez 0; PUSH address "KT1FPmpucXoiX7ZLahj1V1E5tRah1XvcnkZB"; TRANSFER_TOKENS; CONS; PUSH address "KT1PkygK9CqgNLyuJ9iMFcgx1651BrTjN1Q9"; CONTRACT %setRewardRecipient address; IF_NONE { PUSH int 229; FAILWITH } {}; PUSH mutez 0; PUSH address "KT1FPmpucXoiX7ZLahj1V1E5tRah1XvcnkZB"; TRANSFER_TOKENS; CONS; PUSH address "KT1SPUvH5khHtirTEVeECiKrnh4FFXxWZ6ui"; CONTRACT %setRewardRecipient address; IF_NONE { PUSH int 229; FAILWITH } {}; PUSH mutez 0; PUSH address "KT1FPmpucXoiX7ZLahj1V1E5tRah1XvcnkZB"; TRANSFER_TOKENS; CONS; PUSH address "KT1Ad5yJzoiRRdMJPvhJiPJ7Cq8WbJnCS7bg"; CONTRACT %setRewardRecipient address; IF_NONE { PUSH int 229; FAILWITH } {}; PUSH mutez 0; PUSH address "KT1FPmpucXoiX7ZLahj1V1E5tRah1XvcnkZB"; TRANSFER_TOKENS; CONS; PUSH address "KT1WgguedKZWucrdRKQXaRECEPMZennaVPck"; CONTRACT %setRewardRecipient address; IF_NONE { PUSH int 229; FAILWITH } {}; PUSH mutez 0; PUSH address "KT1FPmpucXoiX7ZLahj1V1E5tRah1XvcnkZB"; TRANSFER_TOKENS; CONS; PUSH address "KT1CkpDuwCFrnoqTam6upYiPBiFNsSEVbBei"; CONTRACT %setRewardRecipient address; IF_NONE { PUSH int 229; FAILWITH } {}; PUSH mutez 0; PUSH address "KT1FPmpucXoiX7ZLahj1V1E5tRah1XvcnkZB"; TRANSFER_TOKENS; CONS }
+{ DROP; NIL operation; PUSH address "KT1STLQKxiRtAh1e7DZhu1xUTAJ7KLpV9Rru"; CONTRACT %setTargetOracle address; IF_NONE { PUSH string "Invalid entrypoint: setTargetOracle"; FAILWITH } {}; PUSH mutez 0; PUSH address "KT1Cy1wwHbW1NMawnQCrfXMBtuwXSsidEDg2"; TRANSFER_TOKENS; CONS }
